requirements.txt
supabase>=2.3.0
pandas>=2.1.0
python-dotenv>=1.0.0
